Move configureStore to redux
